% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linesplit.R
\name{linesplit}
\alias{linesplit}
\title{Example Wrapper for 'linesplit' Example}
\usage{
linesplit(str, return_as_arrow = FALSE)
}
\arguments{
\item{str}{A single character value containing several lines separated by newline
characters}

\item{return_as_arrow}{A boolean flag to enable Arrow object return}
}
\value{
The incoming line split and returned as either a 'nanoarrow' array or an
'Arrow' array
}
\description{
Example Wrapper for 'linesplit' Example
}
\examples{
txt <- "the\\nquick\\nbrown\\nfox"
linesplit(txt)
linesplit(txt, TRUE)
}
