\name{linesplitter-package}
\alias{linesplitter-package}
\alias{linesplitter}
\docType{package}
\title{\packageTitle{linesplitter}}
\description{\packageDescription{linesplitter}}
\section{Package Content}{\packageIndices{linesplitter}}
\author{\packageAuthor{linesplitter}}
\section{Maintainer}{\packageMaintainer{linesplitter}}
\keyword{package}
